requirements.txt
coreapi==2.3.3
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0


